Refactor CCS tests (#121547)
* Refactor CCS tests

Clean up duplication, move stop tests to separate test

(cherry picked from commit dbf15393a79e960151975056810b24d9c6e748b1)

# Conflicts:
#	muted-tests.yml
#	x-pack/plugin/esql/src/internalClusterTest/java/org/elasticsearch/xpack/esql/action/CrossClusterAsyncQueryIT.java
#	x-pack/plugin/esql/src/internalClusterTest/java/org/elasticsearch/xpack/esql/action/CrossClusterQueryIT.java
#	x-pack/plugin/esql/src/internalClusterTest/java/org/elasticsearch/xpack/esql/action/CrossClusterQueryUnavailableRemotesIT.java
